requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
